+++ docs/releasenotes.rst
@@ -14,6 +14,8 @@ Next release
 * Two new columns in :ref:`list output <list>`: *arguments* and *info*.
   Can be hidden with: ``mq ls -c aI-``.
 * Deprecated ``venv/activate`` script.  Use ``venv/bin/activate`` instead.
+* Resources can now be specified in the scripts as special comments:
+  ``# MQ: resources=24:2h``.
 
 
 Version 20.9.1